Diabetes Diagnosis Using Logistic Regression and Support Vector Machine (SVM) Algorithms
In this study, Logistic Regression and Support Vector Machine (SVM) algorithms were applied to diagnose diabetes. Logistic Regression is a probability-based model extensively utilized in classification problems. This algorithm calculates the likelihood of an observation belonging to a specific class when the dependent variable is binary. On the other hand, Support Vector Machine is a classification method that aims to determine a hyperplane which optimally separates the data into two classes, proving particularly effective in high-dimensional data scenarios. In this research, both algorithms were tested for their efficacy in diagnosing diabetic patients.